backend/requirements.txt
# =============================================
# Production Dependencies
# =============================================

# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Configuration
python-dotenv==1.0.0
pydantic[email]==2.5.0

# External Services
openai==1.57.4
edge-tts==7.2.3
google-cloud-storage==2.13.0

# =============================================
# Development Dependencies
# =============================================
# Uncomment the following for development environment
# Or install with: pip install -r requirements.txt

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # For testing FastAPI

# Code Quality (optional for production)
# black==23.12.0
# flake8==6.1.0
# isort==5.13.2
# mypy==1.7.1

# Development Tools (optional)
# ipython==8.18.1
# ipdb==0.13.13
